Include emicklei/go-restful in directory move
