+++ instrumentation/github.com/emicklei/go-restful/example/docker-compose.yml
@@ -26,7 +26,7 @@ services:
   go-restful-server:
     build:
       dockerfile: $PWD/Dockerfile
-      context: ../../../..
+      context: ../../../../..
     ports:
       - "8080:80"
     command:
